requirements.txt
fastapi==0.104.1
uvicorn==0.23.2

sqlalchemy==2.0.23
psycopg2-binary==2.9.9

pydantic-settings==2.0.3

python-dotenv==1.0.0

pandas==2.1.3
numpy==1.26.2

httpx==0.25.1

pytest==8.0.0